requirements.txt
gradio>=4.21.0
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.8.4
openai>=1.23.6
python-dotenv>=1.0.1